Can edit own ratings and add short review.  Not read out anywhere, timestamps applied on new data
