+++ newmovieratings/users/forms.py
@@ -2,6 +2,7 @@ from django import forms
 from django.contrib.auth.models import User
 from moviedata.models import Rating
 
+
 class LoginForm(forms.ModelForm):
     class Meta:
         model = User
@@ -11,4 +12,9 @@ class LoginForm(forms.ModelForm):
 class RatingForm(forms.ModelForm):
     class Meta:
         model = Rating
-        fields = ('movie','score')
+        fields = ('movie','score','review')
+
+class UserRatingForm(forms.ModelForm):
+    class Meta:
+        model = Rating
+        fields = ('score','review')
